+++ src/main/java/com/neeraj/SpringEcom/exception/ProductNotFoundException.java
@@ -2,7 +2,7 @@ package com.neeraj.SpringEcom.exception;
 
 public class ProductNotFoundException extends RuntimeException {
 
-    public ProductNotFoundException(int id) {
+    public ProductNotFoundException(Long id) {
         super("Product not found with id: " + id);
     }
